RbfInterpolationTools

Introduction

Documentation for RbfInterpolationTools.

A simple package to perform interpolations using RadialBasisFunctions (RBF).

Note that the package was developed to teach basic computational methods and palying around with Julia.

Context

There are several ways to use mathematical propieties of functions to estimate value of functions without actually having to compute the function, but actually using a set of observed points.

The RBF apporach uses real functions such as Gussian, Multiquadratic, Inverse Multiquadratic and Quadratic to construct a basis to estimate escalar valued functions. Thus, to estimate a desired value at some coordinates, a weighted sum of terms coming from each of the known points and using a function of the distance between them.

It has been shown that using RBF tools yields a good method for the solution of Differential Equations (including PDE) see this work, it is also heavily used in areas such as geographical tools for example, image processing, physical simulations, and many other fields.

First steps

The closing square bracket switches to the package manager interface and the add command installs the package and any missing dependencies. To return to the Julia REPL hit the delete key.

To load the package run

julia> using RbfInterpolationTools

Quick-start

Suppose we have observed some points at random times:

julia> t=hcat(rand(30) .* 6.28)

And the observables (for example, of speed) are:

julia> v_observed = vec(map(x -> sin(x), X))

To estimate the values at those points we are going to create a interpolation environment with:

julia> using RbfInterpolationTools

The sintax for the interpolation environment is simple:

julia> inter = RbfInterpolationTools.Interpolator(X = t, y = y)

Now, we just have to get a Matrix with the points where we want to know the values

julia> new_ts = hcat(0:6.28/(999):6.28)

And simply we get the desired values with:

julia> new_vs = RbfInterpolationTools.interpolateMesh(inter, new_ts)

RbfInterpolationTools.Interpolator — Type

Interpolator environment setup :: mutable struct Interpolator __________________________________________________________________________________________________ Parameters: | X::Matrix Matrix to set interpolation | y::Vector Vector with interpolate ready variables. | ϵ::Float64 Epsilon to set basis function => Default : 1. | basisname::Strin Selected for user ready basis | φ::Function Function of (distance between points , epsilon) => Default : phiMQ(r,ϵ) | w::SVector Static Vector containing the weights | __________________________________________________________________________________________________|

RbfInterpolationTools._build — Method

build the onterpolation matrix

-> Input: :basis: ::Matrix with data to be interpolated ->Output: :Weights vector: Vector containing the weights of the interpolation

RbfInterpolationTools._get_kernel — Function

The function to retrive the basis for interpolation -> Input: :basis: String with the name of the function being selected. The available options are: ["MQ" , "InvMQ" , "Quadratic , "Gaussian"] ->Output: :basis function: two float parameters=> e (the epsilopn of the fuction) r (Distance form center)

RbfInterpolationTools.gradientPoint — Method

Function to interpolate one point's gradient __________________________________________________________________________________________________ Parameters: | inter::Interpolator An interpolation environment already setted up | p::Vector Coordinates of the point to interpolate | Output: | Interpolation results setted by y on Interpolator and using ForwardDiff | __________________________________________________________________________________________________|

RbfInterpolationTools.interpolatePoint — Method

Function to interpolate one point __________________________________________________________________________________________________ Parameters: | inter::Interpolator An interpolation environment already setted up | p::Vector Coordinates of the point to interpolate | Output: | Interpolation results setted by y on Interpolator | __________________________________________________________________________________________________|