RbfInterpolationTools

Documentation for RbfInterpolationTools.

RbfInterpolationTools.Interpolator

RbfInterpolationTools._build

RbfInterpolationTools._get_kernel

RbfInterpolationTools.gradientPoint

RbfInterpolationTools.interpolatePoint

RbfInterpolationTools.Interpolator — Type

Interpolator environment setup :: mutable struct Interpolator __________________________________________________________________________________________________ Parameters: | X::Matrix Matrix to set interpolation | y::Vector Vector with interpolate ready variables. | ϵ::Float64 Epsilon to set basis function => Default : 1. | basisname::Strin Selected for user ready basis | φ::Function Function of (distance between points , epsilon) => Default : phiMQ(r,ϵ) | w::SVector Static Vector containing the weights | __________________________________________________________________________________________________|

RbfInterpolationTools._build — Method

build the onterpolation matrix

-> Input: :basis: ::Matrix with data to be interpolated ->Output: :Weights vector: Vector containing the weights of the interpolation

RbfInterpolationTools._get_kernel — Function

The function to retrive the basis for interpolation -> Input: :basis: String with the name of the function being selected. The available options are: ["MQ" , "InvMQ" , "Quadratic , "Gaussian"] ->Output: :basis function: two float parameters=> e (the epsilopn of the fuction) r (Distance form center)

RbfInterpolationTools.gradientPoint — Method

Function to interpolate one point's gradient __________________________________________________________________________________________________ Parameters: | inter::Interpolator An interpolation environment already setted up | p::Vector Coordinates of the point to interpolate | Output: | Interpolation results setted by y on Interpolator and using ForwardDiff | __________________________________________________________________________________________________|

RbfInterpolationTools.interpolatePoint — Method

Function to interpolate one point __________________________________________________________________________________________________ Parameters: | inter::Interpolator An interpolation environment already setted up | p::Vector Coordinates of the point to interpolate | Output: | Interpolation results setted by y on Interpolator | __________________________________________________________________________________________________|